PhoneBook/requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.1.4
streamlit==1.29.0
